unique-id: add module and generator; db-repo: use int64 as model id;
The unique id generator provides you with a uint64 k-sortable unique id based on snowflake / sonyflake id.
It's backed by a dynamo db table for an instance identifier to prevent collisions.
To be able to store these unique ids in the database, you have to use 64 bit integers.
For consistency, this is enforced on all models.

+++ pkg/db-repo/errors.go
@@ -6,12 +6,12 @@ import (
 )
 
 type RecordNotFoundError struct {
-	id      uint
+	id      int64
 	modelId string
 	err     error
 }
 
-func NewRecordNotFoundError(id uint, modelId string, err error) RecordNotFoundError {
+func NewRecordNotFoundError(id int64, modelId string, err error) RecordNotFoundError {
 	return RecordNotFoundError{
 		id:      id,
 		modelId: modelId,
